Interpreter (#18)
* Create exceptions for interpreter

* Create interpreter prototype

* Add interpretation to main file

* Add F.iml to .gitignore

* Fix ListASTNode

* Improve ReturnASTNode

* Create default function handlers

* Improve DefaultFunctionHandler

* Implement default boolean functions

* Improve arithmetic default functions

* Change long to int

* Implement equal, nonequal

* Change long to int

* Implement some default functions

* Implement some default functions

* Create template files

* Create template files

* Add functions to mapper

* Add default functions handling to ListASTNode

* Add raw exceptions

* Improve ListASTNode

* Fix bug

* Fix sample

* Fix sample

* Improve ListASTNode

* Fix CondASTNode

* Fix DefaultFunctionMapper

* Fix WhileASTNode

* Fix sample

* Improve ProgASTNode

* Fix sample

* Fix sample

* Fix sample

* Fix DefaultFunctionMapper

* Fix sample

* Implement IsNullDefaultFunctionHandler

* Implement IsIntDefaultFunctionHandler

* Implement IsBoolDefaultFunctionHandler

* Implement IsRealDefaultFunctionHandler

* Implement IsListDefaultFunctionHandler

* Implement IsAtomDefaultFunctionHandler

* Implement eval

* Implement head

* Implement tail

* Implement cons

* Improve LiteralASTNode

* Fix sample

* Fix sample

* Add capturing

* Delete unnecessary file

+++ src/lexer/LexicalAnalysisError.java
@@ -1,7 +1,7 @@
 package lexer;
 
 public class LexicalAnalysisError extends RuntimeException {
-    public LexicalAnalysisError(char c, long line, long column) {
+    public LexicalAnalysisError(char c, int line, int column) {
         super("Lexical analysis error: unexpected character '" + c + "' at line " + line + ", column " + column);
     }
 }
